Split configuration between templates and twig
Updated factory to pull configuration from both the `templates` key as
well as a `twig` top-level key, merging them such that the latter has
precedence over the former.

The structure of the `twig` key is:

```php
return [
    'twig' => [
        'cache_dir' => 'path to cached templates',
        'assets_url' => 'base URL for assets',
        'assets_version' => 'base version for assets',
        'extensions' => [
            // extension service names or instances
        ],
    ],
];
```

Note that the `cache_dir` and `assets_*` values have moved to this
array; they are still honored if in the `templates` array, but they
should be migrated.

Also note that the verbiage `helpers` has been replaced with
`extensions` to keep the verbiage consistent with the renderer.